refactor(repl): move the completer into pkg/complete
Test files are not allowed at the repo root, so the completer needs a home
next to its test. It depends only on rt and vm — readline's AutoCompleter
is satisfied structurally, so the package does not pull readline in.
