+++ lg_repl.go
@@ -12,57 +12,12 @@ import (
 	"os"
 	"path/filepath"
 	"strings"
-	"unicode"
 
 	"github.com/chzyer/readline"
 	"github.com/nooga/let-go/pkg/compiler"
-	"github.com/nooga/let-go/pkg/rt"
+	"github.com/nooga/let-go/pkg/complete"
 	"github.com/nooga/let-go/pkg/vm"
 )
-
-func isCompletionTerminator(r rune) bool {
-	switch r {
-	case '(', ')', '[', ']', '{', '}', '"', '\\', '\'', '@', '`', '~', ';', '#':
-		return true
-	}
-	return unicode.IsSpace(r)
-}
-
-// completer implements readline's AutoCompleter interface.
-// readline passes line as []rune and pos as a rune index; we stay in runes
-// throughout so non-ASCII input and mid-line cursors are handled correctly.
-type completer struct {
-	ctx *compiler.Context
-}
-
-func (c *completer) Do(line []rune, pos int) (newLine [][]rune, length int) {
-	if pos > len(line) {
-		pos = len(line)
-	}
-	head := line[:pos]
-
-	start := pos
-	for start > 0 && !isCompletionTerminator(head[start-1]) {
-		start--
-	}
-	prefix := string(head[start:])
-
-	symbols := rt.FuzzyNamespacedSymbolLookup(c.ctx.CurrentNS(), vm.Symbol(prefix))
-
-	// readline INSERTS the candidate at the cursor, so a candidate has to be the
-	// part not typed yet — returning whole symbols turns "ns-unm<tab>" into
-	// "ns-unmns-unmap". Candidates come back unqualified (the lookup resolves the
-	// ns segment of the prefix and matches on the name), so the typed part to
-	// strip is the name segment, not the whole word. length is the same segment's
-	// rune count: readline echoes that many runes ahead of each candidate when it
-	// lists them.
-	_, namePrefix, _ := vm.Symbol(prefix).NamespacedRaw()
-	for _, s := range symbols {
-		newLine = append(newLine, []rune(strings.TrimPrefix(string(s), string(namePrefix))+" "))
-	}
-	length = len([]rune(string(namePrefix)))
-	return
-}
 
 // ANSI color codes for syntax highlighting. ansiReset and ansiBold are
 // declared in lg_ansi.go; the rest are repl-only.
@@ -135,7 +90,7 @@ func repl(ctx *compiler.Context) {
 	config := &readline.Config{
 		Prompt:          prompt,
 		EOFPrompt:       "exit",
-		AutoComplete:    &completer{ctx: ctx},
+		AutoComplete:    &complete.Completer{Ctx: ctx},
 		InterruptPrompt: "^C",
 		HistoryFile:     historyFile(),
 		Painter:         &syntaxHighlighter{},
